Refactor routes into modular routers
